requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
pymongo==4.4.0
langdetect==1.0.9
requests==2.31.0
openai>=1.0.0
flask-limiter>=3.5.0
